backups/object-recover/requirements.txt
psycopg2-binary>=2.9,<3.0
pyyaml>=5.4,<6.0
minio==7.1.3
